# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/api/v1/coupon/create" : {
      "post" : {
        "summary" : "创建优惠券",
        "description" : "创建优惠券",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "name" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "description" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "type" : {
                    "type" : "integer",
                    "description" : "类型(1-常驻,2-秒抢)"
                  },
                  "faceValue" : {
                    "type" : "number",
                    "description" : ""
                  },
                  "minSpend" : {
                    "type" : "number",
                    "description" : ""
                  },
                  "validDays" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "perUserLimit" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "totalStock" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "seckillTotalStock" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "remainingStock" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "seckillRemainingStock" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "status" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "createTime" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "updateTime" : {
                    "type" : "string",
                    "description" : ""
                  }
                },
                "description" : "优惠券信息"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "状态码，0 表示成功"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "id" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "name" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "description" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "type" : {
                          "type" : "integer",
                          "description" : "类型(1-常驻,2-秒抢)"
                        },
                        "faceValue" : {
                          "type" : "number",
                          "description" : ""
                        },
                        "minSpend" : {
                          "type" : "number",
                          "description" : ""
                        },
                        "validDays" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "perUserLimit" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "totalStock" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "seckillTotalStock" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "remainingStock" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "seckillRemainingStock" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "status" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "createTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "updateTime" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : "返回数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/coupon/{id}" : {
      "get" : {
        "summary" : "根据ID查询优惠券",
        "description" : "根据ID查询优惠券",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "优惠券ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "状态码，0 表示成功"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "id" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "name" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "description" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "type" : {
                          "type" : "integer",
                          "description" : "类型(1-常驻,2-秒抢)"
                        },
                        "faceValue" : {
                          "type" : "number",
                          "description" : ""
                        },
                        "minSpend" : {
                          "type" : "number",
                          "description" : ""
                        },
                        "validDays" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "perUserLimit" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "totalStock" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "seckillTotalStock" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "remainingStock" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "seckillRemainingStock" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "status" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "createTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "updateTime" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : "返回数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/coupon/admin/grant" : {
      "post" : {
        "summary" : "后台管理接口：批量发放优惠券",
        "description" : "后台管理接口：批量发放优惠券",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "description" : "用户ID列表",
                "items" : {
                  "type" : "integer"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "状态码，0 表示成功"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "boolean",
                      "description" : "返回数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}